+++ client-go/applyconfiguration/core/v1beta1/internalpackagemanagerspec.go
@@ -13,25 +13,45 @@ import (
 // InternalPackageManagerSpecApplyConfiguration represents a declarative configuration of the InternalPackageManagerSpec type for use
 // with apply.
 type InternalPackageManagerSpecApplyConfiguration struct {
-	License          *product.LicenseSpec                `json:"license,omitempty"`
-	Volume           *product.VolumeSpec                 `json:"volume,omitempty"`
-	NodeSelector     map[string]string                   `json:"nodeSelector,omitempty"`
-	AddEnv           map[string]string                   `json:"addEnv,omitempty"`
-	Image            *string                             `json:"image,omitempty"`
-	ImagePullPolicy  *v1.PullPolicy                      `json:"imagePullPolicy,omitempty"`
-	S3Bucket         *string                             `json:"s3Bucket,omitempty"`
-	Replicas         *int                                `json:"replicas,omitempty"`
-	DomainPrefix     *string                             `json:"domainPrefix,omitempty"`
-	BaseDomain       *string                             `json:"baseDomain,omitempty"`
-	GitSSHKeys       []SSHKeyConfigApplyConfiguration    `json:"gitSSHKeys,omitempty"`
-	AzureFiles       *AzureFilesConfigApplyConfiguration `json:"azureFiles,omitempty"`
-	AdditionalConfig *string                             `json:"additionalConfig,omitempty"`
+	Enabled             *bool                               `json:"enabled,omitempty"`
+	Teardown            *bool                               `json:"teardown,omitempty"`
+	License             *product.LicenseSpec                `json:"license,omitempty"`
+	Volume              *product.VolumeSpec                 `json:"volume,omitempty"`
+	NodeSelector        map[string]string                   `json:"nodeSelector,omitempty"`
+	AddEnv              map[string]string                   `json:"addEnv,omitempty"`
+	Image               *string                             `json:"image,omitempty"`
+	ImagePullPolicy     *v1.PullPolicy                      `json:"imagePullPolicy,omitempty"`
+	S3Bucket            *string                             `json:"s3Bucket,omitempty"`
+	Replicas            *int                                `json:"replicas,omitempty"`
+	DomainPrefix        *string                             `json:"domainPrefix,omitempty"`
+	BaseDomain          *string                             `json:"baseDomain,omitempty"`
+	GitSSHKeys          []SSHKeyConfigApplyConfiguration    `json:"gitSSHKeys,omitempty"`
+	AzureFiles          *AzureFilesConfigApplyConfiguration `json:"azureFiles,omitempty"`
+	AdditionalConfig    *string                             `json:"additionalConfig,omitempty"`
+	Auth                *AuthSpecApplyConfiguration         `json:"auth,omitempty"`
+	OIDCClientSecretKey *string                             `json:"oidcClientSecretKey,omitempty"`
 }
 
 // InternalPackageManagerSpecApplyConfiguration constructs a declarative configuration of the InternalPackageManagerSpec type for use with
 // apply.
 func InternalPackageManagerSpec() *InternalPackageManagerSpecApplyConfiguration {
 	return &InternalPackageManagerSpecApplyConfiguration{}
+}
+
+// WithEnabled sets the Enabled field in the declarative configuration to the given value
+// and returns the receiver, so that objects can be built by chaining "With" function invocations.
+// If called multiple times, the Enabled field is set to the value of the last call.
+func (b *InternalPackageManagerSpecApplyConfiguration) WithEnabled(value bool) *InternalPackageManagerSpecApplyConfiguration {
+	b.Enabled = &value
+	return b
+}
+
+// WithTeardown sets the Teardown field in the declarative configuration to the given value
+// and returns the receiver, so that objects can be built by chaining "With" function invocations.
+// If called multiple times, the Teardown field is set to the value of the last call.
+func (b *InternalPackageManagerSpecApplyConfiguration) WithTeardown(value bool) *InternalPackageManagerSpecApplyConfiguration {
+	b.Teardown = &value
+	return b
 }
 
 // WithLicense sets the License field in the declarative configuration to the given value
@@ -152,5 +172,21 @@ func (b *InternalPackageManagerSpecApplyConfiguration) WithAzureFiles(value *Azu
 // If called multiple times, the AdditionalConfig field is set to the value of the last call.
 func (b *InternalPackageManagerSpecApplyConfiguration) WithAdditionalConfig(value string) *InternalPackageManagerSpecApplyConfiguration {
 	b.AdditionalConfig = &value
+	return b
+}
+
+// WithAuth sets the Auth field in the declarative configuration to the given value
+// and returns the receiver, so that objects can be built by chaining "With" function invocations.
+// If called multiple times, the Auth field is set to the value of the last call.
+func (b *InternalPackageManagerSpecApplyConfiguration) WithAuth(value *AuthSpecApplyConfiguration) *InternalPackageManagerSpecApplyConfiguration {
+	b.Auth = value
+	return b
+}
+
+// WithOIDCClientSecretKey sets the OIDCClientSecretKey field in the declarative configuration to the given value
+// and returns the receiver, so that objects can be built by chaining "With" function invocations.
+// If called multiple times, the OIDCClientSecretKey field is set to the value of the last call.
+func (b *InternalPackageManagerSpecApplyConfiguration) WithOIDCClientSecretKey(value string) *InternalPackageManagerSpecApplyConfiguration {
+	b.OIDCClientSecretKey = &value
 	return b
 }